+++ apps/demo-dtr-app/src/App.tsx
@@ -1,3 +1,21 @@
+/**
+ * Copyright (c) 2024, WSO2 LLC. (http://www.wso2.com).
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
 import './App.css'
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import { AuthProvider } from './components/AuthProvider';
@@ -7,6 +25,8 @@ import { ExpandedContextProvider } from "./utils/expanded_context.tsx";
 import { PersistGate } from "redux-persist/integration/react";
 import { store, persistor } from "./redux/store.ts";
 import LoginPage from './pages/login_v2';
+import MissingParamPage from './pages/missing_param';
+import NavBar from './components/NavBar';
 
 // Extend the Window interface to include the Config property
 declare global {
@@ -19,6 +39,7 @@ declare global {
       questionnaire_response: string;
       patient: string;
       medicationRequest: string;
+      ehr_baseUrl: string;
     };
   }
 }
@@ -30,10 +51,35 @@ function App() {
         <ExpandedContextProvider>
           <BrowserRouter>
             <AuthProvider>
-              <Routes>
-                <Route path="/" element={<DrugPiorAuthPage />} />
-                <Route path="/login" element={<LoginPage />} />
-              </Routes>
+              <div style={{ width: "100vw", height: "100vh", backgroundColor: "#f0f0f0" }}>
+                <NavBar />
+                <div
+                  style={{
+                    width: "80vw",
+                    margin: "auto",
+                    marginTop: "30px",
+                    border: "1px solid #ccc",
+                    borderRadius: "8px",
+                    overflow: "hidden",
+                    boxShadow: "0 4px 8px rgba(0, 0, 0, 0.1)",
+                  }}
+                >
+                  <div
+                    style={{
+                      width: "100%",
+                      height: "85vh",
+                      overflow: "auto",
+                      backgroundColor: "#f9f9f9",
+                    }}
+                  >
+                    <Routes>
+                      <Route path="/" element={<DrugPiorAuthPage />} />
+                      <Route path="/login" element={<LoginPage />} />
+                      <Route path="/invalid-req" element={<MissingParamPage />} />
+                    </Routes>
+                  </div>
+                </div>
+              </div>
             </AuthProvider>
           </BrowserRouter>
         </ExpandedContextProvider>
@@ -42,4 +88,4 @@ function App() {
   )
 }
 
-export default App
+export default App;
