+++ apps/demo-dtr-app/src/components/AuthProvider.tsx
@@ -1,3 +1,19 @@
+// Copyright (c) 2024-2025, WSO2 LLC. (http://www.wso2.com).
+//
+// WSO2 LLC. licenses this file to you under the Apache License,
+// Version 2.0 (the "License"); you may not use this file except
+// in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
 import {
   createContext,
   useContext,
@@ -34,15 +50,18 @@ export const AuthProvider = ({ children }: { children: ReactNode }) => {
       navigate(redirectTo, { replace: true });
     } else if (response.status == 401) {
       setIsAuthenticated(false);
-      const coverageId = query.get("coverageId") || "";
-      const medicationRequestId = query.get("medicationRequestId") || "";
-      const patientId = query.get("patientId") || "";
+      const coverageId = query.get("coverageId");
+      const medicationRequestId = query.get("medicationRequestId");
+      const patientId = query.get("patientId");
+
+      if (!coverageId || !medicationRequestId || !patientId) {
+        navigate("/invalid-req");
+        return;
+      }
 
       sessionStorage.setItem("coverageId", coverageId);
       sessionStorage.setItem("medicationRequestId", medicationRequestId);
       sessionStorage.setItem("patientId", patientId);
-
-      console.log("Patient ID (Auth): ", patientId);
 
       navigate("/login");
     }
